Additions: run_sts_system.py, which will run a STS we have developed. For now, it only runs AveragingWord2Vec, which is our basic strategy. To use run_sts_system, enter as the first argument the string representing the algorithm to use (for AveragingWord2Vec, this is 'Averaging'), and then enter as the rest of the arguments the train directories you want to use to train on, e.g., 'STS2012-en-test'. DO NOT use 'STS2012-en-train', as we will use this for our evaluation (i.e., run_sts_system will treat this as the training set always, you just specify which directories you want to use).
That being said, if you just execute run_sts_system with no arguments, it will assume the algorithm to use as 'Averaging', and assume you want to use all the STS201X-en-test directories under the train directory. If you only specify the first argument string - the algorithm - then it will assume only that you want to use all the STS201X-en-test directories.

To add an algorithm to run_sts_system, import your implementation at the top and then scroll to the bottom and add an elif for your algorithm. Inside it, add all the code you need to run your new system (which is a separate python file), and then add a line using the save_results function to save your computed scores to a txt file.

Note that run_sts_system assembles ALL the training data into the variables training_set and training_labels, and all the testing data into testing_set and testing_labels. Pass these to your code for training and testing. The sets look like a list of lists, where the the outer list is the list of sentence pairs, and the inner lists are the pairs of sentences themselves. The labels are simply a list of floating point numbers that match up to the sets of sentences.

NOTE: I added to downloads.sh the necessary code to download the pretrained Google model. Unfortunately, the link doesn't seem to work with curl, but it works if you do it manually. Not sure how to resolve this problem.